requirements.txt
dj-database-url
python-dotenv
psycopg2-binary
django-cors-headers
Django
djangorestframework
